feat(camera): fetch pose azimuths from platform API
credentials.json no longer carries azimuths. A background loop fetches
GET /api/v1/poses/ with each camera token and maps patrol_id to azimuth,
retrying until the API is reachable. A legacy azimuths list in
credentials.json still takes precedence.

+++ pyro_camera_api/pyro_camera_api/camera/adapters/linovision.py
@@ -47,7 +47,7 @@ class LinovisionCamera(BaseCamera, PTZMixin, FocusMixin):
         password: str,
         cam_type: str = "ptz",
         cam_poses: Optional[List[int]] = None,
-        cam_azimuths: Optional[List[int]] = None,
+        cam_azimuths: Optional[List[float]] = None,
         protocol: str = "http",
         verify_tls: bool = False,
         snapshot_channel: str = "101",
